rework for wasm support
